rpc: remove outdated prototype grpc service
