+++ crates/sui-rpc-api/src/service/mod.rs
@@ -1,10 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub(crate) mod checkpoints;
 mod committee;
 pub(crate) mod health;
-mod info;
-pub(crate) mod objects;
 mod protocol_config;
 pub(crate) mod transactions;
